backend/requirements.txt
# FastAPI core
fastapi==0.115.2
uvicorn[standard]==0.30.6
starlette==0.38.2

# Database & ORM
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.13.3

# Security & Auth
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Env & settings
pydantic-settings==2.6.1

# Utilities
python-multipart==0.0.9

# (Optional, if you want rate limiting or file uploads later)
# slowapi==0.1.9
# boto3==1.35.21
